Fikser sammenslåing av søknader med kunPleietrengendeDataFraAndreSøkere.
- Henter alle søknader på følgende spesifikt barn.
- Tester at kun informasjon om omsorgstilbud slås sammen når det er søkander fra flere søkere.
