+++ src/main/kotlin/no/nav/sifinnsynapi/soknad/SøknadRepository.kt
@@ -11,10 +11,7 @@ interface SøknadRepository : JpaRepository<PsbSøknadDAO, String> {
 
     @Query(
         nativeQuery = true,
-        value = "SELECT * FROM psb_søknad WHERE søker_aktør_id = ?1 AND pleietrengende_aktør_id = ?2 ORDER BY oppdatert_dato ASC"
+        value = "SELECT * FROM psb_søknad WHERE pleietrengende_aktør_id = ?1 ORDER BY oppdatert_dato ASC"
     )
-    fun hentSøknaderSortertPåOppdatertTidspunkt(
-        søkerAktørId: String,
-        pleietrengendeAktørIder: String
-    ): Stream<PsbSøknadDAO>
+    fun hentSøknaderPåPleietrengendeSortertPåOppdatertTidspunkt(pleietrengendeAktørIder: String): Stream<PsbSøknadDAO>
 }
