+++ src/test/kotlin/no/nav/sifinnsynapi/soknad/SøknadServiceTest.kt
@@ -1,5 +1,6 @@
 package no.nav.sifinnsynapi.soknad
 
+import assertk.assertions.isEqualTo
 import com.ninjasquad.springmockk.MockkBean
 import io.mockk.every
 import no.nav.k9.søknad.Søknad
@@ -10,6 +11,7 @@ import no.nav.k9.søknad.ytelse.psb.v1.arbeidstid.ArbeidstidInfo
 import no.nav.k9.søknad.ytelse.psb.v1.arbeidstid.ArbeidstidPeriodeInfo
 import no.nav.k9.søknad.ytelse.psb.v1.tilsyn.TilsynPeriodeInfo
 import no.nav.security.token.support.spring.test.EnableMockOAuth2Server
+import no.nav.sifinnsynapi.konsument.k9sak.OnpremKafkaHendelseKonsumentIntegrasjonsTest
 import no.nav.sifinnsynapi.omsorg.OmsorgDAO
 import no.nav.sifinnsynapi.omsorg.OmsorgRepository
 import no.nav.sifinnsynapi.oppslag.BarnOppslagDTO
@@ -17,6 +19,7 @@ import no.nav.sifinnsynapi.oppslag.OppslagsService
 import no.nav.sifinnsynapi.oppslag.SøkerOppslagRespons
 import no.nav.sifinnsynapi.utils.*
 import org.assertj.core.api.Assertions.assertThat
+import org.junit.Assert
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
@@ -42,78 +45,40 @@ import java.util.stream.Stream
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.MOCK) // Integrasjonstest - Kjører opp hele Spring Context med alle konfigurerte beans.
 internal class SøknadServiceTest {
 
-    @MockkBean
+    @Autowired
     private lateinit var søknadRepository: SøknadRepository
-
-    @MockkBean
-    private lateinit var oppslagsService: OppslagsService
 
     @Autowired
     private lateinit var søknadService: SøknadService
 
-    @Autowired
-    lateinit var omsorgRepository: OmsorgRepository
-
     private companion object {
+        private val annenSøkerAktørId = "00000000000"
         private val hovedSøkerAktørId = "11111111111"
         private val barn1AktørId = "22222222222"
-        private val barn2AktørId = "33333333333"
-    }
-
-    @BeforeAll
-    internal fun setUp() {
-        omsorgRepository.saveAll(
-            listOf(
-                OmsorgDAO(
-                    id = "1",
-                    søkerAktørId = hovedSøkerAktørId,
-                    pleietrengendeAktørId = barn1AktørId,
-                    harOmsorgen = true,
-                    opprettetDato = ZonedDateTime.now(UTC),
-                    oppdatertDato = ZonedDateTime.now(UTC)
-                ),
-                OmsorgDAO(
-                    id = "2",
-                    søkerAktørId = hovedSøkerAktørId,
-                    pleietrengendeAktørId = barn2AktørId,
-                    harOmsorgen = true,
-                    opprettetDato = ZonedDateTime.now(UTC),
-                    oppdatertDato = ZonedDateTime.now(UTC)
-                )
-            )
-        )
-    }
-
-    @BeforeEach
-    internal fun beforeEach() {
-        every { oppslagsService.hentAktørId() } returns SøkerOppslagRespons(aktør_id = hovedSøkerAktørId)
-        every { oppslagsService.hentBarn() } returns listOf(
-            BarnOppslagDTO(
-                aktør_id = barn1AktørId,
-                fødselsdato = LocalDate.parse("2005-02-12"),
-                fornavn = "Ole",
-                mellomnavn = null,
-                etternavn = "Doffen",
-                identitetsnummer = "12020567099"
-            ),
-            BarnOppslagDTO(
-                aktør_id = barn2AktørId,
-                fødselsdato = LocalDate.parse("2005-10-30"),
-                fornavn = "Dole",
-                mellomnavn = null,
-                etternavn = "Doffen",
-                identitetsnummer = "30100577255"
-            )
-        )
     }
 
     @Test
-    fun `kan slå sammen perioder med tilsyn`() {
-        every { søknadRepository.hentSøknaderSortertPåOppdatertTidspunkt(any(), any()) } answers {
-            Stream.of(
+    fun `gitt søknader med tilsyn og arbeidstid fra flere søkere på samme barn, forvent kun at tilsynsperioder blir slått sammen`() {
+        val organisasjonsnummer = "987654321"
+
+        // gitt at det eksiterer to søknader med arbeidstid og omsorgstilbud fra 2 søkere på samme barn...
+        søknadRepository.saveAll(
+            listOf(
                 psbSøknadDAO(
                     journalpostId = "1",
+                    søkerAktørId = hovedSøkerAktørId,
+                    pleietrengendeAktørId = barn1AktørId,
                     søknad = defaultSøknad(
+                        arbeidstid = Arbeidstid().medArbeidstaker(
+                            listOf(
+                                defaultArbeidstaker(
+                                    organisasjonsnummer = organisasjonsnummer,
+                                    periode = Periode(LocalDate.parse("2021-08-01"), LocalDate.parse("2021-10-11")),
+                                    normaltTimerPerDag = 8,
+                                    faktiskArbeidTimerPerDag = 4
+                                )
+                            )
+                        ),
                         tilsynsordning = defaultTilsynsordning(
                             mapOf(
                                 Periode(
@@ -128,7 +93,19 @@ internal class SøknadServiceTest {
                 ),
                 psbSøknadDAO(
                     journalpostId = "2",
+                    søkerAktørId = annenSøkerAktørId,
+                    pleietrengendeAktørId = barn1AktørId,
                     søknad = defaultSøknad(
+                        arbeidstid = Arbeidstid().medArbeidstaker(
+                            listOf(
+                                defaultArbeidstaker(
+                                    organisasjonsnummer = "933333333",
+                                    periode = Periode(LocalDate.parse("2021-09-25"), LocalDate.parse("2021-12-01")),
+                                    normaltTimerPerDag = 8,
+                                    faktiskArbeidTimerPerDag = 2
+                                )
+                            )
+                        ),
                         tilsynsordning = defaultTilsynsordning(
                             mapOf(
                                 Periode(
@@ -142,298 +119,38 @@ internal class SøknadServiceTest {
                     )
                 )
             )
-        }
+        )
 
-        val sammenSlåttTilsynsordning =
-            søknadService.hentSøknadsopplysningerPerBarn()
-                .first().søknad.getYtelse<PleiepengerSyktBarn>().tilsynsordning
+        // forent at søknadene blir slått sammen...
+        val søknad: Søknad? = søknadService.slåSammenSøknaderFor(hovedSøkerAktørId, barn1AktørId)
+        Assert.assertNotNull(søknad)
 
+        val ytelse = søknad!!.getYtelse<PleiepengerSyktBarn>()
+        val sammenslåttTilsynsordning = ytelse.tilsynsordning
+        val sammenslåttArbeidstid = ytelse.arbeidstid
+
+        // forvent det kun er en arbeidstaker i den sammenslåtte søknaden...
+        assertThat(sammenslåttArbeidstid.arbeidstakerList.size).isEqualTo(1)
+
+        // og at periodene for arbeidstid kun tilhørerer søker som har gjort oppslaget.
         assertResultet(
-            faktiskePerioder = sammenSlåttTilsynsordning.perioder,
+            faktiskArbeidstaker = sammenslåttArbeidstid.arbeidstakerList.first(),
+            forventetOrganisasjonsnummer = organisasjonsnummer,
+            forventedePerioder = mapOf(
+                Periode(LocalDate.parse("2021-08-01"), LocalDate.parse("2021-10-11")) to ArbeidstidPeriodeInfo()
+                    .medFaktiskArbeidTimerPerDag(Duration.ofHours(4))
+                    .medJobberNormaltTimerPerDag(Duration.ofHours(8))
+            )
+        )
+
+        // Samt, at periodene med omsorgstilbd fra de to søknadene med forskjellige søkere er slått sammen.
+        assertResultet(
+            faktiskePerioder = sammenslåttTilsynsordning.perioder,
             forventedePerioder = mapOf(
                 Periode(LocalDate.parse("2021-08-01"), LocalDate.parse("2021-09-24")) to
                         TilsynPeriodeInfo().medEtablertTilsynTimerPerDag(Duration.ofHours(4)),
                 Periode(LocalDate.parse("2021-09-25"), LocalDate.parse("2021-12-01")) to
                         TilsynPeriodeInfo().medEtablertTilsynTimerPerDag(Duration.ofHours(2).plusMinutes(30))
-            )
-        )
-    }
-
-    @Test
-    fun `kan slå sammen arbeidstid for en arbeidstaker`() {
-        val org = "987654321"
-        every { søknadRepository.hentSøknaderSortertPåOppdatertTidspunkt(any(), any()) } answers {
-            Stream.of(
-                psbSøknadDAO(
-                    journalpostId = "1",
-                    søknad = defaultSøknad(
-                        arbeidstid = Arbeidstid().medArbeidstaker(
-                            listOf(
-                                defaultArbeidstaker(
-                                    organisasjonsnummer = org,
-                                    periode = Periode(LocalDate.parse("2021-08-01"), LocalDate.parse("2021-10-11")),
-                                    normaltTimerPerDag = 8,
-                                    faktiskArbeidTimerPerDag = 4
-                                )
-                            )
-                        )
-                    )
-                ),
-                psbSøknadDAO(
-                    journalpostId = "2",
-                    søknad = defaultSøknad(
-                        arbeidstid = Arbeidstid().medArbeidstaker(
-                            listOf(
-                                defaultArbeidstaker(
-                                    organisasjonsnummer = org,
-                                    periode = Periode(LocalDate.parse("2021-09-25"), LocalDate.parse("2021-12-01")),
-                                    normaltTimerPerDag = 8,
-                                    faktiskArbeidTimerPerDag = 2
-                                )
-                            )
-                        )
-                    ),
-                )
-            )
-        }
-
-        val sammenslåttArbeidstid =
-            søknadService.hentSøknadsopplysningerPerBarn().first().søknad.getYtelse<PleiepengerSyktBarn>().arbeidstid
-        assertThat(sammenslåttArbeidstid.arbeidstakerList.size).isEqualTo(1)
-        assertResultet(
-            faktiskArbeidstaker = sammenslåttArbeidstid.arbeidstakerList.first(),
-            forventetOrganisasjonsnummer = org,
-            forventedePerioder = mapOf(
-                Periode(LocalDate.parse("2021-08-01"), LocalDate.parse("2021-09-24")) to ArbeidstidPeriodeInfo()
-                    .medFaktiskArbeidTimerPerDag(Duration.ofHours(4))
-                    .medJobberNormaltTimerPerDag(Duration.ofHours(8)),
-
-                Periode(LocalDate.parse("2021-09-25"), LocalDate.parse("2021-12-01")) to ArbeidstidPeriodeInfo()
-                    .medFaktiskArbeidTimerPerDag(Duration.ofHours(2))
-                    .medJobberNormaltTimerPerDag(Duration.ofHours(8))
-            )
-        )
-    }
-
-    @Test
-    fun `kan slå sammen arbeidstid for flere arbeidstakere`() {
-        val org1 = "911111111";
-        val org2 = "922222222";
-        val org3 = "933333333";
-        val org4 = "944444444";
-        every { søknadRepository.hentSøknaderSortertPåOppdatertTidspunkt(any(), any()) } answers {
-            Stream.of(
-                psbSøknadDAO(
-                    journalpostId = "1",
-                    søknad = defaultSøknad(
-                        arbeidstid = Arbeidstid().medArbeidstaker(
-                            listOf(
-                                defaultArbeidstaker(
-                                    organisasjonsnummer = org1,
-                                    periode = Periode(LocalDate.parse("2021-08-01"), LocalDate.parse("2021-10-11")),
-                                    normaltTimerPerDag = 8,
-                                    faktiskArbeidTimerPerDag = 4
-                                ),
-                                defaultArbeidstaker(
-                                    organisasjonsnummer = org2,
-                                    periode = Periode(LocalDate.parse("2021-08-05"), LocalDate.parse("2021-09-12")),
-                                    normaltTimerPerDag = 8,
-                                    faktiskArbeidTimerPerDag = 4
-                                ),
-                                defaultArbeidstaker(
-                                    organisasjonsnummer = org3,
-                                    periode = Periode(LocalDate.parse("2021-06-01"), LocalDate.parse("2021-06-10")),
-                                    normaltTimerPerDag = 8,
-                                    faktiskArbeidTimerPerDag = 3
-                                )
-                            )
-                        )
-                    )
-                ),
-                psbSøknadDAO(
-                    journalpostId = "2",
-                    søknad = defaultSøknad(
-                        arbeidstid = Arbeidstid().medArbeidstaker(
-                            listOf(
-                                defaultArbeidstaker(
-                                    organisasjonsnummer = org1,
-                                    periode = Periode(LocalDate.parse("2021-09-25"), LocalDate.parse("2021-12-01")),
-                                    normaltTimerPerDag = 8,
-                                    faktiskArbeidTimerPerDag = 2
-                                ),
-                                defaultArbeidstaker(
-                                    organisasjonsnummer = org2,
-                                    periode = Periode(LocalDate.parse("2021-09-25"), LocalDate.parse("2021-12-01")),
-                                    normaltTimerPerDag = 8,
-                                    faktiskArbeidTimerPerDag = 2
-                                ),
-                                defaultArbeidstaker(
-                                    organisasjonsnummer = org4,
-                                    periode = Periode(LocalDate.parse("2021-11-13"), LocalDate.parse("2021-11-15")),
-                                    normaltTimerPerDag = 8,
-                                    faktiskArbeidTimerPerDag = 2
-                                )
-                            )
-                        )
-                    ),
-                )
-            )
-        }
-
-        val sammenslåttYtelse =
-            søknadService.hentSøknadsopplysningerPerBarn().first().søknad.getYtelse<PleiepengerSyktBarn>()
-        val resultatArbeidstakere = sortertArbeidstakere(sammenslåttYtelse)
-        assertThat(resultatArbeidstakere.size).isEqualTo(4)
-
-        assertResultet(
-            faktiskArbeidstaker = resultatArbeidstakere[0],
-            forventetOrganisasjonsnummer = org1,
-            forventedePerioder = mapOf(
-                Periode(LocalDate.parse("2021-08-01"), LocalDate.parse("2021-09-24")) to ArbeidstidPeriodeInfo()
-                    .medFaktiskArbeidTimerPerDag(Duration.ofHours(4))
-                    .medJobberNormaltTimerPerDag(Duration.ofHours(8)),
-                Periode(LocalDate.parse("2021-09-25"), LocalDate.parse("2021-12-01")) to ArbeidstidPeriodeInfo()
-                    .medFaktiskArbeidTimerPerDag(Duration.ofHours(2))
-                    .medJobberNormaltTimerPerDag(Duration.ofHours(8))
-            )
-        )
-
-        assertResultet(
-            faktiskArbeidstaker = resultatArbeidstakere[1],
-            forventetOrganisasjonsnummer = org2,
-            forventedePerioder = mapOf(
-                Periode(LocalDate.parse("2021-08-05"), LocalDate.parse("2021-09-12")) to ArbeidstidPeriodeInfo()
-                    .medFaktiskArbeidTimerPerDag(Duration.ofHours(4))
-                    .medJobberNormaltTimerPerDag(Duration.ofHours(8)),
-                Periode(LocalDate.parse("2021-09-25"), LocalDate.parse("2021-12-01")) to ArbeidstidPeriodeInfo()
-                    .medFaktiskArbeidTimerPerDag(Duration.ofHours(2))
-                    .medJobberNormaltTimerPerDag(Duration.ofHours(8))
-            )
-        )
-
-        assertResultet(
-            faktiskArbeidstaker = resultatArbeidstakere[2],
-            forventetOrganisasjonsnummer = org3,
-            forventedePerioder = mapOf(
-                Periode(LocalDate.parse("2021-06-01"), LocalDate.parse("2021-06-10")) to ArbeidstidPeriodeInfo()
-                    .medFaktiskArbeidTimerPerDag(Duration.ofHours(3))
-                    .medJobberNormaltTimerPerDag(Duration.ofHours(8))
-            )
-        )
-
-        assertResultet(
-            faktiskArbeidstaker = resultatArbeidstakere[3],
-            forventetOrganisasjonsnummer = org4,
-            forventedePerioder = mapOf(
-                Periode(LocalDate.parse("2021-11-13"), LocalDate.parse("2021-11-15")) to ArbeidstidPeriodeInfo()
-                    .medFaktiskArbeidTimerPerDag(Duration.ofHours(2))
-                    .medJobberNormaltTimerPerDag(Duration.ofHours(8))
-            )
-        )
-    }
-
-    @Test
-    fun `kan slå sammen arbeidstid for frilanser`() {
-        every { søknadRepository.hentSøknaderSortertPåOppdatertTidspunkt(any(), any()) } answers {
-            Stream.of(
-                psbSøknadDAO(
-                    journalpostId = "1",
-                    søknad = defaultSøknad(
-                        arbeidstid = Arbeidstid().medFrilanserArbeidstid(
-                            ArbeidstidInfo().medPerioder(
-                                mapOf(
-                                    Periode(LocalDate.parse("2021-08-01"), LocalDate.parse("2021-10-11")) to
-                                            ArbeidstidPeriodeInfo()
-                                                .medJobberNormaltTimerPerDag(Duration.ofHours(8))
-                                                .medFaktiskArbeidTimerPerDag(Duration.ofHours(4))
-                                )
-                            )
-                        )
-                    )
-                ),
-                psbSøknadDAO(
-                    journalpostId = "2",
-                    søknad = defaultSøknad(
-                        arbeidstid = Arbeidstid().medFrilanserArbeidstid(
-                            ArbeidstidInfo().medPerioder(
-                                mapOf(
-                                    Periode(LocalDate.parse("2021-09-25"), LocalDate.parse("2021-12-01")) to
-                                            ArbeidstidPeriodeInfo()
-                                                .medJobberNormaltTimerPerDag(Duration.ofHours(8))
-                                                .medFaktiskArbeidTimerPerDag(Duration.ofHours(2))
-                                )
-                            )
-                        )
-                    ),
-                )
-            )
-        }
-
-        val sammenslåttArbeidstid =
-            søknadService.hentSøknadsopplysningerPerBarn().first().søknad.getYtelse<PleiepengerSyktBarn>().arbeidstid
-        assertResultet(
-            faktiskePerioder = sammenslåttArbeidstid.frilanserArbeidstidInfo.get().perioder,
-            forventedePerioder = mapOf(
-                Periode(LocalDate.parse("2021-08-01"), LocalDate.parse("2021-09-24")) to ArbeidstidPeriodeInfo()
-                    .medFaktiskArbeidTimerPerDag(Duration.ofHours(4))
-                    .medJobberNormaltTimerPerDag(Duration.ofHours(8)),
-
-                Periode(LocalDate.parse("2021-09-25"), LocalDate.parse("2021-12-01")) to ArbeidstidPeriodeInfo()
-                    .medFaktiskArbeidTimerPerDag(Duration.ofHours(2))
-                    .medJobberNormaltTimerPerDag(Duration.ofHours(8))
-            )
-        )
-    }
-
-    @Test
-    fun `kan slå sammen arbeidstid for selvstendig næringsdrivende`() {
-        every { søknadRepository.hentSøknaderSortertPåOppdatertTidspunkt(any(), any()) } answers {
-            Stream.of(
-                psbSøknadDAO(
-                    journalpostId = "1",
-                    søknad = defaultSøknad(
-                        arbeidstid = Arbeidstid().medSelvstendigNæringsdrivendeArbeidstidInfo(
-                            ArbeidstidInfo().medPerioder(
-                                mapOf(
-                                    Periode(LocalDate.parse("2021-08-01"), LocalDate.parse("2021-10-11")) to
-                                            ArbeidstidPeriodeInfo()
-                                                .medJobberNormaltTimerPerDag(Duration.ofHours(8))
-                                                .medFaktiskArbeidTimerPerDag(Duration.ofHours(4))
-                                )
-                            )
-                        )
-                    )
-                ),
-                psbSøknadDAO(
-                    journalpostId = "2",
-                    søknad = defaultSøknad(
-                        arbeidstid = Arbeidstid().medSelvstendigNæringsdrivendeArbeidstidInfo(
-                            ArbeidstidInfo().medPerioder(
-                                mapOf(
-                                    Periode(LocalDate.parse("2021-09-25"), LocalDate.parse("2021-12-01")) to
-                                            ArbeidstidPeriodeInfo()
-                                                .medJobberNormaltTimerPerDag(Duration.ofHours(8))
-                                                .medFaktiskArbeidTimerPerDag(Duration.ofHours(2))
-                                )
-                            )
-                        )
-                    ),
-                )
-            )
-        }
-
-        val sammenslåttArbeidstid =
-            søknadService.hentSøknadsopplysningerPerBarn().first().søknad.getYtelse<PleiepengerSyktBarn>().arbeidstid
-        assertResultet(
-            faktiskePerioder = sammenslåttArbeidstid.selvstendigNæringsdrivendeArbeidstidInfo.get().perioder,
-            forventedePerioder = mapOf(
-                Periode(LocalDate.parse("2021-08-01"), LocalDate.parse("2021-09-24")) to ArbeidstidPeriodeInfo()
-                    .medFaktiskArbeidTimerPerDag(Duration.ofHours(4))
-                    .medJobberNormaltTimerPerDag(Duration.ofHours(8)),
-
-                Periode(LocalDate.parse("2021-09-25"), LocalDate.parse("2021-12-01")) to ArbeidstidPeriodeInfo()
-                    .medFaktiskArbeidTimerPerDag(Duration.ofHours(2))
-                    .medJobberNormaltTimerPerDag(Duration.ofHours(8))
             )
         )
     }
